chore: moved FileSize component from helpers to components
ChangeLog: None

Ticket: None
