+++ frontend/src/js/components/releases/dialogs/fileinformation.js
@@ -17,7 +17,7 @@ import { Delete as DeleteIcon, InsertDriveFile as InsertDriveFileIcon } from '@m
 import { Divider, IconButton } from '@mui/material';
 import { makeStyles } from 'tss-react/mui';
 
-import { FileSize } from '../../../helpers';
+import FileSize from '../../common/filesize';
 import { HELPTOOLTIPS, MenderHelpTooltip } from '../../helptips/helptooltips';
 
 const useStyles = makeStyles()(theme => ({
